Update translation.txt
The way the ○○ nicknames work is that different girls will use a different honorific for the ○○ depending on their personality. So for example, せい○○/Sei-○○ might be Sei-kun for one personality, Sei-san for another, Sei-tan for a third, etc. etc. Basically think of the ○○ as a personality-specific honorific, but it'll ALWAYS be an honorific of some sort.