Whole bunch of updated translations. (#6)
* Update translation.txt

Not sure what 灯台 is though without better context.

* Update translation.txt

Seems to be generic "start of the morning" lines.

* Update translation.txt

That's a yawn at the start.

* Forgot to strip comments.

* Update translation.txt

Not 100% sure where these "Configuration List" entries are defined... if that can get figured out, these can get refined, if need be.

* Update translation.txt

Not 100% sure where these "Configuration List" entries are defined... if that can get figured out, these can get refined, if need be.

* Update translation.txt

"Guide" seems to not be selectable as a personality type, but she is listed in the audio settings. Either a fixed character or a future personality, perhaps?

* Update translation.txt

The way the ○○ nicknames work is that different girls will use a different honorific for the ○○ depending on their personality. So for example, せい○○/Sei-○○ might be Sei-kun for one personality, Sei-san for another, Sei-tan for a third, etc. etc. Basically think of the ○○ as a personality-specific honorific, but it'll ALWAYS be an honorific of some sort.

* Update translation.txt

One new position.

* Update translation.txt

New ones added. Guessing on that "Net" one though, but given the proximity to Tennis Table poses, I'm guessing it's a Volleyball net or something.

* Update translation.txt

Some of them are pretty much total guesses though... refine once these can be found (in Free H or otherwise).

* Update translation.txt

* Update translation.txt

* Update translation.txt

Clarification...

* Update translation.txt

* Update translation.txt

* Update translation.txt

Doing "Body Bag" literally has quite the unfortunate connotation in English...

* Update translation.txt

* Update translation.txt

* Update translation.txt

* Update translation.txt

* Update translation.txt

* Update translation.txt

* Update translation.txt

* Update translation.txt

* Update translation.txt

* Update translation.txt

* Update translation.txt

* Update translation.txt

* Update translation.txt

* Update translation.txt

* Update translation.txt

* Update translation.txt

* Update translation.txt

* Update translation.txt

* Update translation.txt

* Update translation.txt

Yeah it's not literal, but what are you gonna do...?

* Update translation.txt

* Update translation.txt

* Update translation.txt

* Update translation.txt

* Update translation.txt

* Update translation.txt

* Update translation.txt

* Update translation.txt

* Update translation.txt

* Update translation.txt

* Update translation.txt

* Update translation.txt

* Update translation.txt

* Update translation.txt

* Update translation.txt

* Update translation.txt

* Update translation.txt

* Update translation.txt

* Update translation.txt

* Update translation.txt

* Update translation.txt

* Update translation.txt

* Update translation.txt

* Update translation.txt

* Update translation.txt

* Update translation.txt

* Update translation.txt

* Update translation.txt

* Update translation.txt

* Update translation.txt

* Update translation.txt

* Update translation.txt

* Update translation.txt

* Update translation.txt

* Update translation.txt

* Update translation.txt

* Update translation.txt

* Update translation.txt

* Update translation.txt

* Update translation.txt

* Update translation.txt

* Update translation.txt

* Update translation.txt

* Update translation.txt

* Update translation.txt

* Update translation.txt

* Update translation.txt

* Update translation.txt

* Update translation.txt

* Update DayTimeChange.txt

* Update ActionChangeUI.txt

* Update UIAL_ListCtrl.txt

* Update OutputAnmInfo.txt

* Update TimeUIControl.txt

* Update ClassCallConfirmViewer.txt

This is a bit of a guess and a bit of a kludge, but I'm presuming this is roughly how the sentence gets constructed, and I need to make it sound good in English without shifting the actual place around...

* Update CategoryInfo.txt

* Update Comparer.txt

* Update Game.txt

* Update CharaData.txt

* Update Heroine.txt

* Update NetworkDefine.txt

Granted, unless Illusion is not geoblocking non-Japanese IPs or a homebrew character uploader/downloader is created, this is all pretty useless...

* Update Passport.txt

* Update translation.txt

* Fix merge conflicts

Co-authored-by: ManlyMarco <39247311+ManlyMarco@users.noreply.github.com>